+++ src/transform/normalize.rs
@@ -8,7 +8,7 @@ pub struct NormalizeTransform;
 
 impl Transform for NormalizeTransform {
     fn apply(&self, column: &mut Column<Numeric>, _parameters: &Option<HashMap<&str,Numeric>>) {
-        let sum: f32 = column.values().sum();
+        let sum: Numeric = column.values().sum();
         let count = column.values().count();
 
         if count == 0 {
